Moving REST endpoint to REST Metadata
